Refactor downloadJvm out of SharedOptions
